the first attempt to split project into modules
