+++ ml_project/enities/__init__.py
@@ -1,0 +1,2 @@
+from .features import Features
+from .read_config import TrainingParams, read_training_params, fix_path, fix_config
