+++ ml_project/enities/features.py
@@ -7,4 +7,3 @@ class Features:
     categorical_features: List[str]
     numerical_features: List[str]
     target_col: str
-    # use_log_trick: bool = field(default=True)
