+++ ml_project/train.py
@@ -1,34 +1,11 @@
 import pandas as pd
 import pickle
-import joblib
 from sklearn.model_selection import train_test_split
-from sklearn.preprocessing import StandardScaler
 from sklearn.linear_model import LogisticRegression
 from sklearn import metrics
 
-from read_config import read_training_params, TrainingParams
-from fix_path import fix_path, fix_config
-
-
-def preprocess_train_data(data: pd.DataFrame, params: TrainingParams) -> pd.DataFrame:
-    cat_features = params.features.categorical_features
-    num_features = params.features.numerical_features
-
-    df_train_cat = data[cat_features]
-    df_train_num = data[num_features]
-
-    cat_columns = df_train_cat.columns
-    num_columns = df_train_num.columns
-
-    df_cat = pd.get_dummies(df_train_cat, columns=cat_columns, prefix=cat_columns)
-
-    scaler = StandardScaler()
-    transformed_num = scaler.fit_transform(df_train_num.to_numpy())
-    df_num = pd.DataFrame(transformed_num, columns=num_columns)
-    with open(params.scaler_path, "wb") as f:
-        joblib.dump(scaler, f)
-
-    return pd.concat([df_cat, df_num], axis=1)
+from enities import read_training_params, fix_path, fix_config
+from features import preprocess_train_data
 
 
 def train_model(config_path: str):
